app/requirements.txt
flask
flask_cors
flask_sqlalchemy
pymysql